store/encrypt: Change back nonce to block number.
Using random numbers yields (obviously, Im dumb) a different output on each
encryption of the file - which is not desirable, since the hash needs to
stay the same. Using the hash value of a block is not unique, and pseudo-random
seeded by the key or something in the header has no advantage to the block
number, since new blocks in the middle will get the same nonce as the block
that was in that place before.
(and introduces the chance of getting the same number twice).