Implement batch writing to the database.
Now every modification to the database is first cached in memory.
If an error occurs over or after multiple writes, the application
can call batch.Rollback() to forget the last changes.
On success, batch.Flush() should be called to write all modifications
in one match (this should reduce the likeliness of inconsistent disk state)