Implement a first draft of explicit pins
Now pins created by brig are implicit by default (shown as cyan pin),
while pins created by the user (via 'brig pin') are explicit (shown
with a magenta pin sign). The special semantics are:

- If a file is pinned explicit, it will stay like that, even
  when brig pins it internally.
- Explicit pins need to be unpinned explictly. Pin does not
  automatically unpin those.

Further work needs to be done refinement and additional handling
commands to see a list of explicit pins and what commits their in.
Also the unpin operation should be capable of unpinning old state,
without the need to go back.

Note that this is all based on a hack in ipfs, which I better
do not explain right now...