catfs: add `backend` hash member to Node.
A new member called `backend` was added to each Node.
It saves the hash under which we can reach the file in the storage
backend. This is what `content` was previously. Now `content` contains
the hash of the plaintext data before it is encrypted. This should
be used to compare files with same content but different encryption.