fuse: proper file size for copy operations
The old day was wrong since we modified data in place and were changing
file.size. However to add a file, we need to remove it first.
Consequently, modifications of it makes wrong size at removal, which
lead to file system inconsistency. The data was there but, directories
size were reported wrong.

New way, is to store data in array in the file handle and work with data
in memory before staging it at the end. This way is inspired by
bolt-mount implementation.

Also, the file size is correct now when data is overwritten by a shorter
file.