Speed up of IsCached method for a given root hash with ipfs backend
Before IsCached execution time was proportional to number of children
links in the cache (about 4 children per 1 MB of file).
We checked if hash in the store with a separate call to the ipfs server,
and this very time consuming since we have to create network socket for each call.
The check can be done much faster if results of 'ipfs refs local' are cached.
Which ever hash/ref is in that list is guaranteed to be cached by the ipfs.

The downside, the method is a bit simplified and does not check for
presence of a child of the child.