The interaction with ipfs backend is cached and thus much faster now
Operations which ask about availability of a hast at the ipfs backend
are now cached. This lead to major speed up of the IsCached(hash) function.
Especially, on the second call for the same hash. But even the first
time call is faster now, since we cache all local ipfs refs,
i.e. output of 'ipfs refs local'.

We currently cache
- is hash and its children cached
- output of 'ipfs refs local' to check the status of a hash alone (no children)
- children links (hashes) of a hash
The last one gives the main speedup, since check for children is quite slow.
Having the cache of all available hashes is also good. The very first
one, can be done fast without caching if the last two are cached, but
it is a nice bonus.

Caching is done with https://github.com/patrickmn/go-cache package v2.1.0
It has some overkill features for our needs, but it does the job.