build: Use `task` instead of `mage` as build system
Mage's idea of writing Go code to build software is cool, but it's
annoying to use, since `go` won't separate it from the actual app
sources. This often leads to weird errors while building mage itself,
although the error is somewhere in brig.

I now switched to `task` which is easier to use and also clever enough
not to re-build things when not needed. See scripts/install-task.sh
Docs need to be changed and probably TravisCI fails also.