repo: enable unix sockets and refactor daemon start
Previously we used a TCP port, now we can configure a URL.  On Unix
systems the URL points to a unix socket, on other a tcp based URL is
used. This is still work in progress and does not have yet the desired
performance properties.  Therefore the normal stage will still use File
IO for now.

Potential next step is to clean up the password handling.  Maybe even to
remove it completely and only add it again later if we find use for it.
The current code is sprinkled all over the place and make changes harder
and repository structuring harder. A more general approach could be to
enable the easy use of encfs over a built-in helper.

Also a next step is to switch Cat() and Tar() to a stream based
approach. Since both go over TCP sockets already I do not expect any
notable performance issues here. Maybe even a slim chance on
improvement.