Massive redesign to avoid use of MaxUint64 as indicator of unknown size
Addresses #53
Not sure I like the changes, quite a lot of utilities assumed and worked
with uint64 integers. But this is no go if we want to use negative size
as indicator of an error or "do not use me" indicator.

Looks like we reducing size capacity by 2 (we are not using the sign bit).
I understand that if we using 2^63 file size, then we are in far future,
but I fill guilty by putting factor of 2 reduction.

What I like the least, I had replace uint64 with int64 in mio/stream

But if we OK with this, then changes pass all tests and ready to go.