fuse avoiding call to cfs.Stat(path) before we sure the Xattribute exist
Apparently, when WritebackCache is enable, the OS is quite spammy with
requests to Getxattr for attribute "security.capability". Old code use
to call cfs.Stat even if such attribute is not the one which we can
provide. The call cost time and decrease througput by about 10MB/s.

Now we check if we can even provide an attribute, and only then fetch
relevant info.