bug fix: the response to Xattribute must take as much as needed!!!
Pretty much 1st call always comes with req.Size=0.  Our job to respond with all we have.
If response resp.Xattr does not fit to a specified buffer req.Size, fuse will internally
truncate it (see
https://github.com/bazil/fuse/blob/fb710f7dfd05053a3bc9516dd5a7a8f84ead8aab/fuse.go#L1656
and similar
https://github.com/bazil/fuse/blob/fb710f7dfd05053a3bc9516dd5a7a8f84ead8aab/fuse.go#L1618)
and relay it to the
caller with ERANGE error (buffer is too small). The caller should ask
again with a larger buffer.

From other hand truncation to zero is a disaster, since the caller
assumes that there are no Xattributes.

Note that apparently, even if the call comes with allowed req.Siz=0, internally it
has some small buffer. So our first reply often fits and the call ends
without error.